refactor: 정적 파일 요청과 api 호출 라우팅 분리
- router에서는 정적 파일 요청, api 호출을 구분
- 각 상황에 맞는 router 추가
